+++ projects/igniteui-angular-i18n/src/i18n/BG/date-range-picker-resources.ts
@@ -8,4 +8,8 @@ export const DateRangePickerResourceStringsBG = {
     igx_date_range_picker_date_separator: 'до',
     igx_date_range_picker_done_button: 'Завърши',
     igx_date_range_picker_cancel_button: 'Отмени'
+    igx_date_range_picker_last7Days: 'Последните 7 дни',
+    igx_date_range_picker_currentMonth: 'Текущ месец',
+    igx_date_range_picker_last30Days: 'Последните 30 дни',
+    igx_date_range_picker_yearToDate: 'От началото на годината'
 } satisfies MakeRequired<IDateRangePickerResourceStrings>;