+++ projects/igniteui-angular-i18n/src/i18n/ES/date-range-picker-resources.ts
@@ -8,4 +8,8 @@ export const DateRangePickerResourceStringsES = {
     igx_date_range_picker_date_separator: 'a',
     igx_date_range_picker_done_button: 'Listo',
     igx_date_range_picker_cancel_button: 'Cancelar'
+    igx_date_range_picker_last7Days: 'Últimos 7 días',
+    igx_date_range_picker_currentMonth: 'Mes actual',
+    igx_date_range_picker_last30Days: 'Últimos 30 días',
+    igx_date_range_picker_yearToDate: 'Año hasta la fecha',
 } satisfies MakeRequired<IDateRangePickerResourceStrings>;