+++ projects/igniteui-angular-i18n/src/i18n/FR/date-range-picker-resources.ts
@@ -8,4 +8,8 @@ export const DateRangePickerResourceStringsFR = {
     igx_date_range_picker_date_separator: 'à',
     igx_date_range_picker_done_button: 'Terminée',
     igx_date_range_picker_cancel_button: 'Annuler'
+    igx_date_range_picker_last7Days: '7 derniers jours',
+    igx_date_range_picker_currentMonth: 'Mois en cours',
+    igx_date_range_picker_last30Days: '30 derniers jours',
+    igx_date_range_picker_yearToDate: 'Année à ce jour',
 } satisfies MakeRequired<IDateRangePickerResourceStrings>;