+++ projects/igniteui-angular-i18n/src/i18n/NB/date-range-picker-resources.ts
@@ -8,4 +8,8 @@ export const DateRangePickerResourceStringsNB = {
     igx_date_range_picker_date_separator: 'til',
     igx_date_range_picker_done_button: 'Ferdig',
     igx_date_range_picker_cancel_button: 'Avbryt'
+    igx_date_range_picker_last7Days: 'Siste 7 dager',
+    igx_date_range_picker_currentMonth: 'Denne måneden',
+    igx_date_range_picker_last30Days: 'Siste 30 dager',
+    igx_date_range_picker_yearToDate: 'Året til dato',
 } satisfies MakeRequired<IDateRangePickerResourceStrings>;