+++ projects/igniteui-angular-i18n/src/i18n/PT/date-range-picker-resources.ts
@@ -8,4 +8,8 @@ export const DateRangePickerResourceStringsPT = {
     igx_date_range_picker_date_separator: 'para',
     igx_date_range_picker_done_button: 'Concluído',
     igx_date_range_picker_cancel_button: 'Cancelar'
+    igx_date_range_picker_last7Days: 'Últimos 7 dias',
+    igx_date_range_picker_currentMonth: 'Mês atual',
+    igx_date_range_picker_last30Days: 'Últimos 30 dias',
+    igx_date_range_picker_yearToDate: 'Ano até hoje',
 } satisfies MakeRequired<IDateRangePickerResourceStrings>;