+++ projects/igniteui-angular-i18n/src/i18n/RO/date-range-picker-resources.ts
@@ -8,4 +8,8 @@ export const DateRangePickerResourceStringsRO = {
     igx_date_range_picker_date_separator: 'la',
     igx_date_range_picker_done_button: 'Gata',
     igx_date_range_picker_cancel_button: 'Anulează'
+    igx_date_range_picker_last7Days: 'Ultimele 7 zile',
+    igx_date_range_picker_currentMonth: 'Luna curentă',
+    igx_date_range_picker_last30Days: 'Ultimele 30 de zile',
+    igx_date_range_picker_yearToDate: 'De la începutul anului',
 } satisfies MakeRequired<IDateRangePickerResourceStrings>;