+++ projects/igniteui-angular-i18n/src/i18n/SV/date-range-picker-resources.ts
@@ -8,4 +8,8 @@ export const DateRangePickerResourceStringsSV = {
     igx_date_range_picker_date_separator: 'till',
     igx_date_range_picker_done_button: 'Färdig',
     igx_date_range_picker_cancel_button: 'Avbryt'
+    igx_date_range_picker_last7Days: 'Senaste 7 dagarna',
+    igx_date_range_picker_currentMonth: 'Aktuell månad',
+    igx_date_range_picker_last30Days: 'Senaste 30 dagarna',
+    igx_date_range_picker_yearToDate: 'Året hittills',
 } satisfies MakeRequired<IDateRangePickerResourceStrings>;