+++ projects/igniteui-angular-i18n/src/i18n/TR/date-range-picker-resources.ts
@@ -8,4 +8,8 @@ export const DateRangePickerResourceStringsTR = {
     igx_date_range_picker_date_separator: '-',
     igx_date_range_picker_done_button: 'Bitti',
     igx_date_range_picker_cancel_button: 'İptal'
+    igx_date_range_picker_last7Days: 'Son 7 gün',
+    igx_date_range_picker_currentMonth: 'Geçerli ay',
+    igx_date_range_picker_last30Days: 'Son 30 gün',
+    igx_date_range_picker_yearToDate: 'Yılbaşı itibarıyla',
 } satisfies MakeRequired<IDateRangePickerResourceStrings>;